pyproject.toml
[project]
name = "reolink"
version = "0.2.0"
description = "Interact with Reolink cameras"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.11.11",
    "click>=8.1.8",
    "pillow>=11.1.0",
    "psycopg2>=2.9.10",
    "pydantic<2",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.0.1",
    "sqlalchemy<2",
    "toolz>=1.0.0",
]

[project.entry-points.console_scripts]
reolink = "reolink.cli:cli"

[project.optional-dependencies]
linters = [
    "ruff>=0.9.1",
]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.uv]
package = true
dev-dependencies = [
    "cython>=3.0.11",
    "pytest-asyncio>=0.25.2",
    "pytest>=8.3.4",
    "setuptools>=75.8.0",
]



[tool.setuptools]
ext-modules = [
    { name = "reolink.interval.c_interval", sources = ["src/reolink/interval/c_interval.c"]}
]

[tool.ruff]
target-version = "py310"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I001", "UP", "ANN", "B", "C", "I", "LOG", "G", "RSE", "RET", "SLOT", "SIM", "TC", "INT", "ARG",
    "PTH", "ERA", "PL", "TRY", "FLY", "RUF", "FA"]
ignore = ["ANN401", "ANN204", "ANN201", "ANN202", "ARG002", "TRY003", "C901", "PLR"]

[tool.ruff.lint.flake8-annotations]
ignore-fully-untyped = true
suppress-dummy-args = true

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

